qbert states + commands implemented

+++ QBert/HealthComponent.h
@@ -7,7 +7,7 @@ namespace dae
 	class HealthComponent : public Component
 	{
 	public:
-		HealthComponent(std::shared_ptr<GameObject> game_object, float baseHealth, std::unique_ptr<Subject> subject) : Component(game_object), m_Health{ baseHealth }, m_Subject{ std::move(subject)} {};
+		HealthComponent(std::shared_ptr<GameObject> game_object, int baseHealth, std::unique_ptr<Subject> subject) : Component(game_object), m_Health{ baseHealth }, m_Subject{ std::move(subject)} {};
 		virtual ~HealthComponent() override = default;
 		//Rule of five
 		HealthComponent(const HealthComponent& other) = delete;
@@ -16,14 +16,14 @@ namespace dae
 		HealthComponent& operator=(HealthComponent&& other) = delete;
 
 
-		void TakeDamage(float damage_amount);
+		void TakeDamage(int damage_amount);
 		float GetHealth() { return m_Health; };
 
 		Subject* const GetSubject() const { return m_Subject.get(); }
 
 
 	private:
-		float m_Health;
+		int m_Health;
 		std::unique_ptr<Subject> m_Subject;
 	};
 }
